Struct rusoto_marketplacecommerceanalytics::MarketplaceCommerceAnalyticsClient

pub struct MarketplaceCommerceAnalyticsClient<P = CredentialsProvider, D = RequestDispatcher> where
    P: ProvideAwsCredentials,
    D: DispatchSignedRequest,  { /* fields omitted */ }

A client for the AWS Marketplace Commerce Analytics API.

Methods

impl MarketplaceCommerceAnalyticsClient

pub fn simple(region: Region) -> MarketplaceCommerceAnalyticsClient

Creates a simple client backed by an implicit event loop.

The client will use the default credentials provider and tls client.

See the rusoto_core::reactor module for more details.

impl<P, D> MarketplaceCommerceAnalyticsClient<P, D> where
P: ProvideAwsCredentials,
D: DispatchSignedRequest,

pub fn new(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> Self